shortend tracing
