+++ helios_runtime/src/pipeline/autonomy_pipeline.rs
@@ -13,7 +13,7 @@ use helios_core::{
 };
 
 use crate::{
-    pipeline::{node::HOST_PRODUCER_ID, rate_gate::RateTimer},
+    pipeline::{key_format::format_key_short, node::HOST_PRODUCER_ID, rate_gate::RateTimer},
     port::{ChannelKey, PortBus},
     prelude::{
         AgentRuntime, Health, NodeId, PipelineBuildError, PipelineNode, Stamped, TickContext,
@@ -304,39 +304,6 @@ fn format_keys(keys: &[ChannelKey]) -> String {
     format!("[{joined}]")
 }
 
-/// Render a [`ChannelKey`] with each `::`-qualified path collapsed to its
-/// leaf — `helios_core::data::sensor::SensorReading` → `SensorReading`.
-/// Used only by the DAG-dump for readability; production code keeps full
-/// paths via [`ChannelKey::Display`].
-fn format_key_short(key: &ChannelKey) -> String {
-    let mut out = String::with_capacity(key.type_name.len());
-    let mut segment_start = 0usize;
-    let bytes = key.type_name.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        let c = bytes[i] as char;
-        if matches!(c, '<' | '>' | ',' | ' ' | '&' | '(' | ')') {
-            push_leaf(&mut out, &key.type_name[segment_start..i]);
-            out.push(c);
-            i += 1;
-            segment_start = i;
-        } else {
-            i += 1;
-        }
-    }
-    push_leaf(&mut out, &key.type_name[segment_start..]);
-
-    if key.instance.trim().is_empty() {
-        out
-    } else {
-        format!("{out} @ \"{}\"", key.instance)
-    }
-}
-
-fn push_leaf(out: &mut String, segment: &str) {
-    let leaf = segment.rsplit("::").next().unwrap_or(segment);
-    out.push_str(leaf);
-}
 
 /// A built, validated autonomy pipeline.
 ///
