+++ helios_runtime/src/pipeline/mod.rs
@@ -15,6 +15,7 @@
 pub mod autonomy_pipeline;
 pub mod build_error;
 pub mod builders;
+pub(crate) mod key_format;
 pub mod node;
 pub mod nodes;
 pub mod rate_gate;
